pyproject.toml
[tool.poetry]
name = "django-quickstart"
version = "0.1.0"
description = ""
authors = ["Pedro Lopez <pedrolopz.85@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
celery = "^5.3.6"
psycopg2 = "^2.8.0"
redis = "^5.0.1"
django-redis = "^5.4.0"
pandas = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
